+++ src/main/java/library_project/models/Library.java
@@ -20,7 +20,7 @@ public class Library {
         return allReservations.getAll();
     }
 
-    public int requestBook(int bookId) throws BookIsOutException {
+    public Long requestBook(Long bookId) throws BookIsOutException {
         Reservation newReservation = null;
         synchronized (this) {
             // check that the book is *not* out
@@ -33,7 +33,7 @@ public class Library {
         throw new BookIsOutException();
     }
 
-    public void returnBook(int reservationId) {
+    public void returnBook(Long reservationId) {
         allReservations.remove(reservationId);
     }
 
@@ -90,7 +90,7 @@ public class Library {
     /*
      * Finds all reservations related to bookId
      */
-    private Iterator<Reservation> findReservations(int bookId) {
+    private Iterator<Reservation> findReservations(Long bookId) {
         Repository<Reservation> foundReservations = new ReservationRepository();
         Iterator<Reservation> reservationIterator = allReservations.getAll();
         Reservation currReservation = null;
@@ -107,7 +107,7 @@ public class Library {
         return foundReservations.getAll();
     }
 
-    public boolean isOut(int bookId) {
+    public boolean isOut(Long bookId) {
         Iterator<Reservation> relevantReservations = findReservations(bookId);
         boolean isOut = false;
