+++ src/main/java/library_project/repos/Repository.java
@@ -6,8 +6,8 @@ import java.util.Iterator;
  * Created by pshek on 11/08/2016.
  */
 public interface Repository<E> {
-    E get(int id);
+    E get(Long id);
     Iterator<E> getAll();
     void add(E entity);
-    void remove(int id);
+    void remove(Long id);
 }
